rutas protegiadas

+++ controllers/reviewsC.js
@@ -41,15 +41,17 @@ reviewsC.allReview = (req, res) => {
 reviewsC.delReview = (req, res) => {
   const { id } = req.params;
 
-  Reviews.destroy({ where: { id } })
-    .then((filasAfectadas) => {
-      if (filasAfectadas > 0) {
-        res.sendStatus(200);
-      } else {
-        res.sendStatus(202);
-      }
-    })
-    .catch(() => res.sendStatus(400));
+  if (req.user.admin) {
+    Reviews.destroy({ where: { id } })
+      .then((filasAfectadas) => {
+        if (filasAfectadas > 0) {
+          res.sendStatus(200);
+        } else {
+          res.sendStatus(202);
+        }
+      })
+      .catch(() => res.sendStatus(400));
+  }
 };
 
 module.exports = reviewsC;
